Create README.md
I am an experienced medical speech pathologist with close to 10 years of experience and a master's degree in speech-language pathology.  I work full-time at a large hospital in Columbus, Ohio, and use my free-time to learn data analytics skills, including: Spreadsheets, SQL, Tableau, and Visualizations.  I am in the process of completing the Data Analytics Google Certificate to deepen my knowledge and skills of data analytics.  My goal is to expand my career by leveraging my knowledge and experience of health insurance, medications, medical diagnosis, and medical terminology into data analytics related to healthcare.  

I find vulnerability in continuing education and the workplace to be a top skill, as this is when we ask questions and reach out to those with more knowledge and experience, which in turn, helps us to grow.  I am looking forward to learning from this community.